Adding d35e support in autoTemplates.
This is currently working for Categories and Overrides. Burning hand and
cone are working with no specific configuration, working on the
measureTemplate of the spell item out of the box.

Overrides are also working well (eg. setting water wave on detect
magic). Configuration are loaded with the system kind of
measuredtemplates (circle, cone, ray, cube).

Note that the autoTemplate/d35e wraps the AbilityTemplate static method
fromItem in pure js. The wrapping through libwrapper should also be
added. Working on it...